moved user-listing fetch to another file
