+++ client/src/components/UserListing.js
@@ -1,38 +1,18 @@
 import React, { useEffect, useState } from "react";
+import UserListingTile from "./UserListingTile";
+import getUserListing from "../apiClient/getUserListing";
 
 const UserListing = (props) => {
     const [userListings, setUserListings] = useState([])
 
-    const getUserListing = async () => {
-        try {
-            const response = await fetch("/api/v1/userListings")
-            if (!response.status) {
-                const error = new Error(`${response.status} (${response.statusText})`);
-                throw error;
-            }
-            const responseData = await response.json();
-            setUserListings(responseData.listings);
-        } catch (error) {
-            console.error("Error in fetch!");
-            console.error(error.message);
-        }
-    }
-
     useEffect(() => {
-        getUserListing()
+        getUserListing().then(listings => {
+            setUserListings(listings)
+        })
     }, [])
 
     const listings = userListings.map((listing) => {
-        return (
-            <div key={listing.id}>
-                <li>
-                    <h4>{listing.title} - {listing.price}$</h4>
-                    <p>{listing.description}</p>
-                    {listing.condition &&  <p>Condition - {listing.condition}</p>}
-                    <br></br>
-                </li>
-            </div>
-        );
+        return <UserListingTile listing={listing} />
     });
 
     return (
